Began CR Part of Assignment of Implementing a profile managment tool.
Need to Fix the Image resizing in Profile Managment. Maybe need to fix the JSX HTML tags to acheive that

Need to further complete Profile Managment (Second Part) that allows user to change profile picture, email and passsword.

Need to implement a size limit on the files, to ensure that Local Storage does not get too full.